palindrome-program
hello in this program there are two programmers errors one is a semicolon(;) after the instillation of flag another is calling the palindrome method with the wrong name both are fixed by myself
